feat: 태그 api 분리
